Move RequireAuth component to the routes directory
