+++ src/Root.tsx
@@ -1,5 +1,5 @@
 import { Outlet, Route, Routes } from 'react-router-dom';
-import { RequireAuth } from './components/RequireAuth';
+import { RequireAuth } from './routes/RequireAuth';
 import { Dashboard } from './components/views/Dashboard/Dashboard';
 import { LoginPanel } from './components/views/LoginPanel/LoginPanel';
 import { DefaultRoute } from './routes/DefaultRoute';
